Refactor templates
