+++ Sources/Vehicle/Templates/Actions/CarPlayActionsTemplate+Build.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+extension CarPlayActionsTemplate {
+    static func build() -> any CarPlayTemplateProvider {
+        CarPlayActionsTemplate(viewModel: .init())
+    }
+}
