+++ Sources/Vehicle/Templates/Actions/CarPlayActionsTemplate.swift
@@ -5,23 +5,27 @@ import Shared
 
 @available(iOS 15.0, *)
 final class CarPlayActionsTemplate: CarPlayTemplateProvider {
-    private var actionsToken: NotificationToken?
     private var actions: Results<Action>?
+    private let viewModel: CarPlayActionsViewModel
 
     var template: CPListTemplate
 
     weak var interfaceController: CPInterfaceController?
 
-    init() {
+    init(viewModel: CarPlayActionsViewModel) {
+        self.viewModel = viewModel
         self.template = CPListTemplate(title: L10n.CarPlay.Navigation.Tab.actions, sections: [])
         template.tabTitle = L10n.CarPlay.Navigation.Tab.actions
         template.tabImage = MaterialDesignIcons.lightningBoltIcon.carPlayIcon(color: nil)
         template.tabSystemItem = .more
+
+        self.viewModel.templateProvider = self
+        template.emptyViewSubtitleVariants = [L10n.SettingsDetails.Actions.title]
     }
 
     func templateWillDisappear(template: CPTemplate) {
         if template == self.template {
-            actionsToken?.invalidate()
+            viewModel.invalidateActionsToken()
         }
     }
 
@@ -32,26 +36,11 @@ final class CarPlayActionsTemplate: CarPlayTemplateProvider {
     }
 
     func update() {
-        let actions = Current.realm().objects(Action.self)
-            .sorted(byKeyPath: "Position")
-            .filter("Scene == nil")
-        updateList(for: actions)
+        viewModel.update()
     }
 
     func updateList(for actions: Results<Action>) {
-        actionsToken?.invalidate()
-        actionsToken = actions.observe { [weak self] _ in
-            self?.updateActions(actions: actions)
-        }
-
         template.updateSections([section(actions: actions)])
-        template.emptyViewSubtitleVariants = [L10n.SettingsDetails.Actions.title]
-    }
-
-    private func updateActions(actions: Results<Action>) {
-        template.updateSections([
-            section(actions: actions),
-        ])
     }
 
     private func section(actions: Results<Action>) -> CPListSection {
@@ -63,20 +52,8 @@ final class CarPlayActionsTemplate: CarPlayTemplateProvider {
                 detailText: action.Text,
                 image: materialDesignIcon
             )
-            item.handler = { _, completion in
-                guard let server = Current.servers.server(for: action) else {
-                    completion()
-                    return
-                }
-                Current.api(for: server).HandleAction(actionID: action.ID, source: .CarPlay).pipe { result in
-                    switch result {
-                    case .fulfilled:
-                        break
-                    case let .rejected(error):
-                        Current.Log.info(error)
-                    }
-                    completion()
-                }
+            item.handler = { [weak self] _, completion in
+                self?.viewModel.handleAction(action: action, completion: completion)
             }
             return item
         }
