+++ Sources/Vehicle/Templates/Areas/CarPlayAreasZonesTemplate+Build.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+@available(iOS 16.0, *)
+extension CarPlayAreasZonesTemplate {
+    static func build() -> any CarPlayTemplateProvider {
+        CarPlayAreasZonesTemplate(viewModel: .init())
+    }
+}
