+++ Sources/Vehicle/Templates/Domains/CarPlayDomainsListTemplate+Build.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+@available(iOS 16.0, *)
+extension CarPlayDomainsListTemplate {
+    static func build() -> any CarPlayTemplateProvider {
+        CarPlayDomainsListTemplate(viewModel: .init())
+    }
+}
