+++ Sources/Vehicle/Templates/Servers/CarPlayServerListTemplate+Build.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+@available(iOS 16.0, *)
+extension CarPlayServersListTemplate {
+    static func build() -> any CarPlayTemplateProvider {
+        CarPlayServersListTemplate(viewModel: .init())
+    }
+}
